refactor: split overlay.go to extract network overlay stuff
